Fix MSVC build compatibility and add miscellaneous updates

+++ tests/instructions/io_instructions.cpp
@@ -34,7 +34,7 @@ TEST_F(InstructionTest,
 
     const uint16_t x = 0x0;
     const uint16_t y = 0xA;
-    const uint8_t n = sprite.size();
+    const uint8_t n = static_cast<uint8_t>(sprite.size());
     const Opcode opcode = 0xD000 | (x << 8) | (y << 4) | n;
 
     // Load the sprite data into memory
@@ -71,7 +71,7 @@ TEST_F(InstructionTest,
 
     const uint16_t x = 0x0;
     const uint16_t y = 0xA;
-    const uint8_t n = sprite.size();
+    const uint8_t n = static_cast<uint8_t>(sprite.size());
     const Opcode opcode = 0xD000 | (x << 8) | (y << 4) | n;
 
     // Load the sprite data into memory
@@ -185,7 +185,7 @@ TEST_F(InstructionTest, LD_VX_K_KeyPressed_SetVxToPressedKey) {
     const uint16_t x = 0x0;
     const Opcode opcode = 0xF00A | (x << 8);
 
-    const uint8_t initialPcValue = registers.pc;
+    const uint16_t initialPcValue = registers.pc;
     const uint8_t key = 0xE;
     keypad[key] = true; // Set key E as pressed
